moving parsing functions to util
